+++ common/src/main/java/com/majruszsaccessories/boosters/Horseshoe.java
@@ -20,7 +20,7 @@ public class Horseshoe extends BoosterHandler {
 	public Horseshoe() {
 		super( MajruszsAccessories.HORSESHOE, Horseshoe.class );
 
-		this.add( LuckBonus.create( 1 ) )
+		this.add( LuckBonus.create( 0.75f ) )
 			.add( SkeletonHorsemanDropChance.create() )
 			.add( TradeOffer.create() );
 	}