+++ common/src/main/java/com/majruszsaccessories/boosters/components/LuckBonus.java
@@ -7,7 +7,7 @@ import com.majruszlibrary.math.Range;
 import com.majruszsaccessories.common.AccessoryHolders;
 import com.majruszsaccessories.common.BonusComponent;
 import com.majruszsaccessories.common.BonusHandler;
-import com.majruszsaccessories.config.RangedInteger;
+import com.majruszsaccessories.config.RangedFloat;
 import com.majruszsaccessories.items.BoosterItem;
 import com.majruszsaccessories.tooltip.TooltipHelper;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
@@ -15,17 +15,17 @@ import net.minecraft.world.entity.ai.attributes.Attributes;
 
 public class LuckBonus extends BonusComponent< BoosterItem > {
 	final AttributeHandler attribute;
-	final RangedInteger bonus = new RangedInteger().id( "luck_bonus" );
+	final RangedFloat bonus = new RangedFloat().id( "luck_bonus" );
 
-	public static ISupplier< BoosterItem > create( int luck ) {
+	public static ISupplier< BoosterItem > create( float luck ) {
 		return handler->new LuckBonus( handler, luck );
 	}
 
-	protected LuckBonus( BonusHandler< BoosterItem > handler, int luck ) {
+	protected LuckBonus( BonusHandler< BoosterItem > handler, float luck ) {
 		super( handler );
 
 		this.attribute = new AttributeHandler( "%s_luck_bonus".formatted( handler.getId() ), ()->Attributes.LUCK, AttributeModifier.Operation.ADDITION );
-		this.bonus.set( luck, Range.of( 1, 10 ) );
+		this.bonus.set( luck, Range.of( 0.0f, 10.0f ) );
 
 		OnPlayerTicked.listen( this::updateLuck )
 			.addCondition( Condition.isLogicalServer() )