+++ changelog.md
@@ -4,4 +4,5 @@
 - removed unused code
 - increased Angler Rune luck bonus from 3 to 3.5
 - decreased Dice and Golden Dice bonuses from 20%/40% to 15%/30%
+- decreased Horseshoe and Golden Horseshoe bonuses from +1/+2 to +0.75/+1.5
 - fixed bug with Horseshoe luck bonus being affected by efficiency bonus